pyproject.toml
[tool.poetry]
name = "hello-world-backend"
version = "0.0.0"
description = "Add a short project description here."
authors = ["Hello World <support@hello_world.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "~3.2.16"
django-environ = "^0.9.0"
django-sites = "^0.11"
argon2-cffi = "^21.3"
python-dotenv = "^0.21.0"
django-cors-headers = "^3.13.0"

# Extensions
# -------------------------------------
pytz = "^2022.7.1"

# Models
# -------------------------------------
psycopg2-binary = "^2.9.5"
Pillow = "^9.4.0"
django-extensions = "^3.2.1"
django-uuid-upload-path = "^1.0"
django-versatileimagefield = "^2.2"

# REST APIs
# -------------------------------------
djangorestframework = "^3.14.0"
drf-yasg = "^1.20"

# Documentation
# -------------------------------------
mkdocs = "^1.4.2"

# Documentation Extensions
# -------------------------------------
mkdocs-git-revision-date-localized-plugin = "^1.1.0"
mkdocs-material = "^9.0.5"
markdown-include = "^0.8.0"

# Sentry SDK
# --------------------------
sentry-sdk = "^1.13.0"

# LOGGING
# -------------------------------------
django-log-request-id = "^2.0.0"

# Async Tasks
# -------------------------------------
django-post-request-task = "^0.5"

# Mail Stuff
# -------------------------------------

# Auth Stuff
# -------------------------------------
PyJWT = "^2.6.0"
django-mail-templated = "^2.6.5"
django-rest-knox = "^4.2.0"

# Production settings
# =====================================

# Static Files and Media Storage
# -------------------------------------
django-storages = "^1.13.2"
boto3 = "~1.26.52"

# Caching
# -------------------------------------
django-redis = "^5.2.0"
hiredis = "^2.1.1"



# Logging
# -------------------------------------
newrelic = "~8.5.0"


[tool.poetry.dev-dependencies]
pre-commit = "^2.21.0"

ansible = "~7.1.0"

# Documentation
# -------------------------------------
isort = "^5.11.4"
black = "~23.1a1"
flake8 = "^6.0.0"

# Debugging
# -------------------------------------
django-debug-toolbar = "^3.8.1"
ipdb = "^0.13.11"

# Testing and coverage
# -------------------------------------
pytest = "^7.2.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
django-dynamic-fixture = "^3.1.2"
pytest-mock = "^3.10.0"
mypy = "~0.991"
django-stubs = "^1.13.1"

# Versioning
# -------------------------------------
bump2version = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_first_party = "hello_world"
default_section = "THIRDPARTY"
import_heading_firstparty = "Hello World Stuff"
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third Party Stuff"
skip_glob = ["*/migrations/**", "*/venv/**", "*/docs/**"]

[tool.black]
target_version = [ "py38", "py39" ]
include = "\\.pyi?$"
exclude = """
/(\\.git/
  |\\.eggs
  |\\.hg
  |__pycache__
  |\\.cache
  |\\.ipynb_checkpoints
  |\\.mypy_cache
  |\\.pytest_cache
  |\\.tox
  |venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |templates
  |locale
)/
"""
